Deduplicate RecordingUI
