+++ lib/spack/spack/test/cmd/env.py
@@ -10,13 +10,12 @@ import pathlib
 import shutil
 import sys
 from argparse import Namespace
-from typing import Any, Dict, List, Optional, Tuple
+from typing import Any, Dict, Optional
 
 import pytest
 
 import spack.cmd.env
 import spack.concretize
-import spack.concretize_ui
 import spack.config
 import spack.environment as ev
 import spack.error
@@ -44,6 +43,7 @@ from spack.spec import Spec
 from spack.stage import stage_prefix
 from spack.store import Store
 from spack.test.conftest import RepoBuilder
+from spack.test.utilities import RecordingUI
 from spack.traverse import traverse_nodes
 from spack.util import tty
 from spack.util.executable import Executable
@@ -5073,16 +5073,6 @@ def test_exists_consistent_with_all_environment_names(
 
     listed = "myenv" in ev.all_environment_names()
     assert ev.exists("myenv") == listed
-
-
-class RecordingUI(spack.concretize_ui.ConcretizerUI):
-    """Frontend that records the groups it is notified about, instead of announcing them."""
-
-    def __init__(self) -> None:
-        self.groups: List[Tuple[str, bool]] = []
-
-    def on_group_started(self, *, group: str, is_default: bool) -> None:
-        self.groups.append((group, is_default))
 
 
 def test_concretization_reports_groups(environment_from_manifest):
